rename fixtures.js -> robotStateFixtures.js
