+++ protocol-designer/src/step-generation/test-with-flow/fixtures/commandFixtures.js
@@ -7,6 +7,37 @@ import type {
   TouchTipArgsV3,
   CommandV3 as Command,
 } from '@opentrons/shared-data'
+import type {
+  CommandsAndRobotState,
+  CommandCreatorErrorResponse,
+} from '../../types'
+
+/** Used to wrap command creators in tests, effectively casting their results
+ **  to normal response or error response
+ **/
+export function getSuccessResult(
+  result: CommandsAndRobotState | CommandCreatorErrorResponse
+): CommandsAndRobotState {
+  if (result.errors) {
+    throw new Error(
+      `Expected a successful command creator call but got errors: ${JSON.stringify(
+        result.errors
+      )}`
+    )
+  }
+  return result
+}
+
+export function getErrorResult(
+  result: CommandsAndRobotState | CommandCreatorErrorResponse
+): CommandCreatorErrorResponse {
+  if (!result.errors) {
+    throw new Error(
+      `Expected command creator to return errors but got success result`
+    )
+  }
+  return result
+}
 
 export const replaceTipCommands = (tip: number | string): Array<Command> => [
   dropTipHelper('A1'),
